+++ src/lazyfree.c
@@ -2,6 +2,7 @@
 #include "bio.h"
 #include "functions.h"
 #include "cluster.h"
+#include "module.h"
 
 #include <stdatomic.h>
